Cleanup code to use fiber-local storage to store the current worker and current stack frame.  Remove unnecessary code to align stack allocations, and fix use of CILK_STACKSIZE environment variable to modify the current Cilk stack size.
In this approach, each __cilkrts_stack_frame stores a pointer to the
current fiber header.  This pointer is set from a TLS variable when a
new Cilk frame is entered.  The runtime system also updates this
pointer when the current fiber associated with a __cilkrts_stack_frame
changes, specifically,
- when a worker steals the continuation in the associated function,
  and
- when a nontrivial cilk_sync is performed on the frame.

This approach seems to be slightly faster than storing the current
stack frame in the worker structure and storing the current worker in
the stack frame.  By storing the current stack frame in fiber-local
storage, we can avoid storing an extra pointer in the worker
structure.  As a result, we trivially avoid issues of false sharing
between modifying the tail pointer and updating the current stack
frame pointer.  Meanwhile, we don't need logic in
__cilkrts_leave_frame{_helper} to update the worker pointer in the
parent __cilkrts_stack_frame.